+++ videogame_collection/src/main/java/com/example/videogame_collection/services/GameService.java
@@ -45,15 +45,15 @@ public class GameService {
 			return gamesRepository.getId(name);
 		}
 		
-		public void updateName(long id, String newName) {
+		public void updateName(int id, String newName) {
 			gamesRepository.updateName(id, newName);
 		}
 		
-		public void updateSystem(long id, String newSystem) {
+		public void updateSystem(int id, String newSystem) {
 			gamesRepository.updateSystem(id, newSystem);
 		}
 		
-		public void updateStatus(long id, String newStatus) {
+		public void updateStatus(int id, String newStatus) {
 			gamesRepository.updateStatus(id, newStatus);
 		}
 }
